Configure CMake to include lua script as raw string
